# openapi.yaml
openapi: 3.0.3
info:
  title: Practing E-Commerce
  description: |-
    This is my first project back and documenting in yaml be patient

    _lorem ipsum!_

  version: 1.0.0
servers:
  - url: https://localhost:3000/api/v1
...
paths:
...
  /products:
    get:
      tags:
        - products
      summary: Return product
      description: Return product
      operationId: createUser
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: Error
    post:
      tags:
        - products
      summary: Create product
      description: Create product
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: The product connot be created
  /products/{productId}:
    get:
      tags:
        - products
      summary: Return product by ID
      description: Return product by ID
      parameters:
        - name: productId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: "False"
    put:
      tags:
        - products
      summary: Update product by ID
      description: Update product by ID
      parameters:
        - name: productId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid Product Id
        "401":
          description: Invalid Category
        "402":
          description: Invalid product
        "504":
          description: The product cannot be created!
    delete:
      tags:
        - products
      summary: Delete product by ID
      description: Delete product by ID
      parameters:
        - name: productId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: The product deleted
        "404":
          description: product not found!
        "400":
          description: Error
  /products/get/count:
    get:
      tags:
        - products
      summary: Return product count
      description: Return product count
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: "False"
  /products//get/featured/{count}:
    get:
      tags:
        - products
      summary: Return product count
      description: Return product count
      parameters:
        - name: count
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
            application/xml:
              schema:
                $ref: "#/components/schemas/Product"
        "500":
          description: "False"
...
  /users:
    get:
      tags:
        - user
      summary: Return user list
      description: Return user list
      operationId: createUsersWithListInput
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: "False"
  /users/register:
    post:
      tags:
        - user
      summary: Creates a new user
      description: Creates a new user
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: The user cannot be created!
  /users/{usersId}:
    get:
      tags:
        - user
      summary: Return User order by ID
      description: Return User order by ID
      parameters:
        - name: usersId
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          description: The user with the given ID was not found
    put:
      tags:
        - user
      summary: Update user by ID
      description: Update user by ID
      parameters:
        - name: usersId
          in: path
          description: ID of the order that needs to be modified
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/User"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: The user cannot be created!
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: ""
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: false
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        "400":
          description: The user not found!
        "401":
          description: Password is wrong!
  /user/register:
    post:
      tags:
        - user
      summary: Register user
      description: Register user
      operationId: updateUser
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
          application/xml:
            schema:
              $ref: "#/components/schemas/User"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: User created
        "400":
          description: The user cannot be created
  /user/{userId}:
    delete:
      tags:
        - user
      summary: Delete user
      description: Delete user
      parameters:
        - name: userId
          in: path
          description: ID of the order that needs to be modified
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: The user deleted
        "400":
          description: Error
        "404":
          description: user not found!
  /user/get/count:
    get:
      tags:
        - user
      summary: Return User count
      description: Return User count
      responses:
        "200":
          description: User Count
        "500":
          description: "False"
components:
  schemas:
    Category:
      type: object
      properties:
        name:
          type: string
          example: luis
        icon:
          type: string
          example: heart
        color:
          type: string
          format: hexadecimal
          example: 0B4C5F
      xml:
        name: category
...
    Product:
      required:
        - name
        - description
        - category
        - countInStock
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          example: Product description
        richDescription:
          type: string
          example: Product richDescription
        image:
          type: string
          default: ""
        images:
          type: array
          xml:
            name: user
            wrapped: true
          items:
            default: ""
        brand:
          type: string
          example: Adidas
          default: ""
        price:
          type: number
          example: 100
        category:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        countInStock:
          type: number
          example: between 0 and 255
        rating:
          type: number
          default: 0
        numReviews:
          type: number
          default: 0
        isFeatured:
          type: boolean
          default: false
        dateCreated:
          type: string
          default: Date.now
      xml:
        name: product
    User:
      required:
        - name
        - email
        - passwordHash
        - phone
        - street
        - apartment
        - city
        - zip
        - country
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        passwordHash:
          type: string
        phone:
          type: string
        isAdmin:
          type: boolean
          default: false
        street:
          type: string
        apartment:
          type: string
        city:
          type: string
        zip:
          type: string
        country:
          type: string
      xml:
        name: user
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header